# api/openapi.yaml
openapi: 3.0.3
info:
  title: Homeworldz Internal API
  version: 0.1.0
servers:
  - url: http://localhost:42000
paths:
...
  /api/v1/regions:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    get:
      operationId: discoverRegions
      summary: Discover regions with active leases
      security:
        - ServiceToken: []
      responses:
        "200":
          description: Active regions
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegionList" }
        "401":
          description: A valid service token is required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      operationId: registerRegion
      summary: Register a region and acquire a lease
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRegionRequest" }
      responses:
        "201":
          description: Region registered
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Region" }
        "400":
          description: Invalid registration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: A valid service token is required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Coordinates are already leased
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/users:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    post:
      operationId: createDevelopmentUser
      summary: Create a development user
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          description: Invalid development user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Username is already registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/sessions:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    post:
      operationId: createSession
      summary: Validate development credentials and create a session
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSessionRequest" }
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          description: Invalid login request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Service authentication or credentials failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/presence:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    get:
      operationId: listPresence
      summary: List non-stale online presence
      security:
        - ServiceToken: []
      responses:
        "200":
          description: Online presence
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PresenceList" }
...
  /api/v1/assets:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    post:
      operationId: registerAssetMetadata
      summary: Register an immutable asset origin or verified replica
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterAssetRequest" }
      responses:
        "201":
          description: Asset metadata registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetMetadata" }
        "400":
          description: Invalid asset registration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Asset UUID conflicts with immutable metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Caller-provided request identifier; unsafe values are replaced.
      schema:
        type: string
        maxLength: 128
        pattern: "^[A-Za-z0-9._-]+$"
...
  headers:
    RequestId:
      required: true
      description: Valid caller-provided request identifier or a generated identifier.
      schema:
        type: string
  securitySchemes:
    ServiceToken:
      type: http
      scheme: bearer
      description: Required by service-to-service operations under /api/.
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    RegisterRegionRequest:
      type: object
      required: [name, gridX, gridY, publicEndpoint]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        gridX: { type: integer, minimum: 0 }
        gridY: { type: integer, minimum: 0 }
        publicEndpoint: { type: string, format: uri }
        viewerPort: { type: integer, minimum: 1, maximum: 65535, default: 42002 }
        leaseSeconds: { type: integer, minimum: 10, maximum: 300, default: 60 }
...
    Region:
      type: object
      required: [id, name, gridX, gridY, publicEndpoint, viewerPort, leaseExpiresAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        gridX: { type: integer }
        gridY: { type: integer }
        publicEndpoint: { type: string, format: uri }
        viewerPort: { type: integer, minimum: 1, maximum: 65535 }
        leaseExpiresAt: { type: string, format: date-time }
    RegionList:
      type: object
      required: [regions]
      properties:
        regions:
          type: array
          items: { $ref: "#/components/schemas/Region" }
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 32, pattern: "^[A-Za-z0-9._-]+$" }
        password: { type: string, minLength: 8, maxLength: 128, format: password }
    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        createdAt: { type: string, format: date-time }
    CreateSessionRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
        sessionSeconds: { type: integer, minimum: 300, maximum: 86400, default: 43200 }
    Session:
      type: object
      required: [id, userId, expiresAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time }
        viewerCircuitCode: { type: integer, minimum: 1, maximum: 4294967295 }
        destinationRegionId: { type: string, format: uuid }
...
    Presence:
      type: object
      required: [userId, regionId, lastSeenAt]
      properties:
        userId: { type: string, format: uuid }
        regionId: { type: string, format: uuid }
        lastSeenAt: { type: string, format: date-time }
    PresenceList:
      type: object
      required: [presence]
      properties:
        presence:
          type: array
          items: { $ref: "#/components/schemas/Presence" }
    RegisterAssetRequest:
      type: object
      additionalProperties: false
      required: [id, creatorUserId, sha256, size, endpoint, origin]
      properties:
        id: { type: string, format: uuid }
        creatorUserId: { type: string, format: uuid }
        sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        size: { type: integer, format: int64, minimum: 1 }
        endpoint: { type: string, format: uri }
        origin: { type: boolean }
    AssetLocation:
      type: object
      required: [endpoint, origin, verifiedAt]
      properties:
        endpoint: { type: string, format: uri }
        origin: { type: boolean }
        verifiedAt: { type: string, format: date-time }
    AssetMetadata:
      type: object
      required: [id, creatorUserId, sha256, size, locations]
      properties:
        id: { type: string, format: uuid }
        creatorUserId: { type: string, format: uuid }
        sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        size: { type: integer, format: int64, minimum: 1 }
        locations:
          type: array
          items: { $ref: "#/components/schemas/AssetLocation" }
...